Moved _gamevars fixture to conftest
