+++ tests/test_main.py
@@ -7,7 +7,6 @@ import stat
 import string
 from collections.abc import Generator
 from pathlib import Path
-from typing import TypeAliasType, get_args
 
 import pytest
 import ruamel.yaml
@@ -84,23 +83,6 @@ def _move_data_files() -> Generator[None]:
     backup_path.rename(data_path)
     assert not backup_path.exists(), f"{backup_path} was not renamed"
     assert data_path.is_dir(), f"{data_path} was not created"
-
-
-@pytest.fixture
-def _gamevars() -> None:
-    """Initialize CLASSIC_Main's gamevars global and validate its types."""
-    assert isinstance(CLASSIC_Main.gamevars, dict), "CLASSIC_Main.gamevars should be initialized to dict"
-    assert len(CLASSIC_Main.gamevars) > 0, "CLASSIC_Main.gamevars should contain default values"
-    assert isinstance(CLASSIC_Main.GameID, TypeAliasType), "CLASSIC_Main.GameID type is unexpected"
-    assert (
-        CLASSIC_Main.GameVars.__annotations__["game"] is CLASSIC_Main.GameID
-    ), "CLASSIC_Main.GameVars type is unexpected"
-    game_ids = get_args(CLASSIC_Main.GameVars.__annotations__["game"].__value__)
-    vr_values = get_args(CLASSIC_Main.GameVars.__annotations__["vr"])
-    assert len(game_ids) > 0, "CLASSIC_Main.GameID type is unexpected"
-    assert all(isinstance(g, str) for g in game_ids), "CLASSIC_Main.GameID type is unexpected"
-    assert CLASSIC_Main.gamevars.get("game") in game_ids, "CLASSIC_Main.gamevars['game'] not initialized"
-    assert CLASSIC_Main.gamevars.get("vr") in vr_values, "CLASSIC_Main.gamevars['vr'] not initialized"
 
 
 @pytest.fixture
